GIX-1047: Rename transaction types file
