+++ frontend/src/lib/modals/sns/SwapModal/ParticipateSwapModal.svelte
@@ -21,7 +21,7 @@
     participateInSwap,
   } from "$lib/services/sns.services";
   import { toastsSuccess } from "$lib/stores/toasts.store";
-  import type { NewTransaction } from "$lib/types/transaction.context";
+  import type { NewTransaction } from "$lib/types/transaction";
   import AdditionalInfoForm from "./AdditionalInfoForm.svelte";
   import AdditionalInfoReview from "./AdditionalInfoReview.svelte";
